requirements.txt
numpy==1.21.5
pandas==1.4.4
streamlite
scikit-learn
pip==22.2.2